* Removing stability module now that everything has moved. * Splitting the plotting functions into separate file.
